+++ packages/core/src/walker.ts
@@ -11,6 +11,8 @@ const SOURCE_GLOBS = [
   "**/*.cjs",
   "**/*.css",
   "**/*.scss",
+  "**/*.svelte",
+  "**/*.vue",
 ];
 
 /**